requirements.txt
pandas
openpyxl
python-dotenv
requests
pytesseract